requirements.txt
openai-whisper
yt-dlp
youtube-transcript-api